Rename causal style files to confirm with *.styles.ts (#1342)
